+++ src/oms_sensemaking/resolution/sensemaker.py
@@ -10,14 +10,15 @@ from oms_sdk.generated.generated_graphql_client import (
     CreateRelationshipInput,
     NodeAttributeQuery,
     NodeAttributeSubQuery,
-    NodeNode,
     NodeQuery,
+    NodesNodes,
     RelationshipNodeQuery,
     RelationshipQuery,
     StringQuery,
 )
 from oms_sdk.generated.generated_graphql_client.enums import Confidence
 
+from oms_sensemaking.clients.instances import aac_client
 from oms_sensemaking.config import SETTINGS
 from oms_sensemaking.core.oms_crud import OmsCrudTool
 from oms_sensemaking.core.sensemakers import FindingBase, Sensemaker
@@ -46,6 +47,14 @@ class DupFinding(FindingBase):
 
     def get_acm(self) -> dict:
         return self.acm
+
+
+@dataclass
+class DupNodeAndAttributeAcms:
+    """Represents a Duplicate Node and Matched Attributes"""
+
+    node: NodesNodes
+    attribute_acms: list[dict]
 
 
 class ResolutionSensemaker(Sensemaker):
@@ -106,7 +115,7 @@ class ResolutionSensemaker(Sensemaker):
 
         criterion: list[list[AttributeAttribute]] = self.gather_criteria(attribute)
         if criterion:
-            dups: list[NodeNode] = self.find_duplicates(criterion)
+            dups: list[DupNodeAndAttributeAcms] = self.find_duplicates(criterion)
             dup_findings: list[DupFinding] = self.create_duplicate_findings(attribute, dups)
             results.extend(dup_findings)
 
@@ -125,34 +134,49 @@ class ResolutionSensemaker(Sensemaker):
             return []
         return AttributeCombinations(attribute, self.oms_crud_tool, self.duplicate_object_iris).gather(node_iri)
 
-    def create_duplicate_findings(self, attribute: AttributeAttribute, nodes: list[NodeNode]) -> list[DupFinding]:
+    def create_duplicate_findings(
+        self, attribute: AttributeAttribute, duplicates: list[DupNodeAndAttributeAcms]
+    ) -> list[DupFinding]:
         """
         Create and return duplicate finding objects from matched nodes
 
         :param attribute: Original attribute being matched on
-        :param nodes: Matched nodes
+        :param duplicates: Matched nodes
         :return: List of DupFinding objects of matches
         """
 
         current_node_id = attribute.nodeId
         dups: list[DupFinding] = []
 
-        for node in nodes:
+        print("\n\n\n")
+        print(len(duplicates))
+
+        for duplicate in duplicates:
+            print(duplicate.node.id)
+
             # Ignore the node we're currently looking at
-            if node.id == current_node_id:
+            if duplicate.node.id == current_node_id:
+                print("ignoring")
                 continue
 
-            dup = DupFinding(start_node_id=current_node_id, end_node_id=node.id, acm=node.acm)
+            print("duplicate.attribute_acms: ", len(duplicate.attribute_acms), duplicate.attribute_acms)
+
+            all_acms = [duplicate.node.acm] + duplicate.attribute_acms
+            print("all_acms: ", len(all_acms), all_acms)
+            rolled_up_acm = aac_client.get_acm_rollup([{"ACM": acm} for acm in all_acms])
+            print("rolled_up_acm: ", rolled_up_acm)
+
+            dup = DupFinding(start_node_id=current_node_id, end_node_id=duplicate.node.id, acm=rolled_up_acm)
 
             rel: CreateRelationshipInput = CreateRelationshipInput(
                 name=SETTINGS.resolution_relationship_name,
                 tags=[SETTINGS.resolution_sensemaker_tag],
                 labels=[SETTINGS.sm_inferenced_label, SETTINGS.res_sm_label, self.version_string],
                 startNodeId=current_node_id,
-                endNodeId=node.id,
+                endNodeId=duplicate.node.id,
                 confidence=Confidence.UNKNOWN,
                 sourceId=attribute.sourceId,
-                acm=attribute.acm,
+                acm=rolled_up_acm,
                 objectPropertyIri=SETTINGS.resolution_relationship_iri,
             )
             created_dup_resolution_rel = self.oms_crud_tool.create_relationship(rel)
@@ -160,7 +184,7 @@ class ResolutionSensemaker(Sensemaker):
             dup.atoms_type = AtomsType.RELATIONSHIP
             dups.append(dup)
 
-            LOGGER.info("Resolution Sensemaker found duplicates %s, %s", current_node_id, node.id)
+            LOGGER.info("Resolution Sensemaker found duplicates %s, %s", current_node_id, duplicate.node.id)
 
         return dups
 
@@ -234,22 +258,29 @@ class ResolutionSensemaker(Sensemaker):
 
         return True, class_iri
 
-    def find_duplicates(self, combinations: list[list[AttributeAttribute]]) -> list[NodeNode]:
+    def find_duplicates(self, combinations: list[list[AttributeAttribute]]) -> list[DupNodeAndAttributeAcms]:
         """
         Determine if there are matching objects in OMSB
 
         :param combinations: A list of lists. The inner lists are various groupings and combinations
         of attributes coming from a current node that may match to other nodes' set of attributes
-        :return: List of duplicate nodes
+        :return: List of DupNodeAndAttributeAcms objects
         """
         for group in combinations:
-            node_attribute_subqueries: list[NodeAttributeSubQuery] = [
-                NodeAttributeSubQuery(
-                    attributeIris=[attribute.attributeIri],
-                    attributeValue=StringQuery(equals=attribute.attributeValue),
+            duplicate_attribute_checks = []
+            current_node_attribute_acms = []
+
+            node_attribute_subqueries: list[NodeAttributeSubQuery] = []
+            for attribute in group:
+                duplicate_attribute_checks.append((attribute.attributeIri, attribute.attributeValue))
+                current_node_attribute_acms.append(attribute.acm)
+
+                node_attribute_subqueries.append(
+                    NodeAttributeSubQuery(
+                        attributeIris=[attribute.attributeIri],
+                        attributeValue=StringQuery(equals=attribute.attributeValue),
+                    )
                 )
-                for attribute in group
-            ]
 
             node_attribute_query: NodeAttributeQuery = NodeAttributeQuery(
                 and_=[NodeAttributeQuery(hasMatch=subquery) for subquery in node_attribute_subqueries]
@@ -260,5 +291,25 @@ class ResolutionSensemaker(Sensemaker):
             nodes_response = self.oms_crud_tool.get_nodes(query)
 
             if nodes_response and nodes_response.data:
-                return nodes_response.data
+                duplicates = []
+
+                for node in nodes_response.data:
+                    attribute_acms = []
+
+                    attributes = self.oms_crud_tool.get_node_attribute_by_iri(
+                        node.id, [check[0] for check in duplicate_attribute_checks]
+                    )
+                    print(attributes)
+                    for attribute in attributes:
+                        attribute_check = (attribute.attributeIri, attribute.attributeValue)
+                        if attribute_check in duplicate_attribute_checks:
+                            print("node: ", node.id)
+                            print("attribute_check: ", attribute_check)
+                            print("attribute.id: ", attribute.id)
+                            print("attribute.acm: ", attribute.acm)
+                            attribute_acms.append(attribute.acm)
+
+                    duplicates.append(DupNodeAndAttributeAcms(node, attribute_acms + current_node_attribute_acms))
+
+                return duplicates
         return []